+++ ICD.Connect.UI/Utils/HtmlUtils.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using ICD.Common.Utils;
 
 namespace ICD.Connect.UI.Utils
 {
@@ -24,6 +25,16 @@ namespace ICD.Connect.UI.Utils
 				throw new ArgumentException("Color is null or empty", "color");
 
 			return string.IsNullOrEmpty(text) ? text : string.Format("<font color=\"{0}\">{1}</font>", color, text);
+		}
+
+		/// <summary>
+		/// Replaces the newlines in the given text with line breaks.
+		/// </summary>
+		/// <param name="text"></param>
+		/// <returns></returns>
+		public static string ReplaceNewlines(string text)
+		{
+			return text == null ? null : text.Replace(IcdEnvironment.NewLine, LINE_BREAK);
 		}
 	}
 }